+++ src/autogenerated/ResourcesAutogen.hpp
@@ -39,6 +39,8 @@ public:
         QPixmap modModeDisabled2;
         QPixmap modModeEnabled;
         QPixmap modModeEnabled2;
+        QPixmap replyDark;
+        QPixmap replyThreadDark;
         QPixmap search;
         QPixmap timeout;
         QPixmap trashCan;